requirements.txt
tensorflow
numpy
matplotlib
medmnist
pillow
streamlit
scikit-image
python-dotenv
